Moved all unique plotting functions from CryoAsicEventViewer to the Analysis class, and removed the CryoAsicEventViewer class. It had so much similar infrastructure as the analysis class, so it makes sense to combine them. Also modified functions to conform to the new channelmapping. Also modified plotting functions to all have the same ax=None, show=True arguments as well as return fig,ax objects for overlaying in notebooks.
